feat(extensions/track-peek): extract as a standalone extension
