+++ src/components/injected/main.css
@@ -3,8 +3,6 @@
 }
 
 :root[data-theme='light'] {
-  --main-song-preview-color: #eaf2ec;
-
   .main-home-content {
     --carousel-start-chevron-gradient: #fafafab3;
     --carousel-end-chevron-gradient: #fafafab3;
@@ -434,19 +432,6 @@
   }
 }
 
-:root[data-theme='dark'] {
-  --main-song-preview-color: #051b0c;
-}
-
-:root[style*='--turntable-play-state: paused'] {
-  .main-nowPlayingBar-center .song-preview-attacher {
-    img,
-    svg {
-      animation: unset;
-    }
-  }
-}
-
 /* Remove Upgrade Button, AD */
 .main-topBar-UpgradeButton,
 [data-testid='test-ref-div'],
@@ -519,52 +504,6 @@
 
 .main-nowPlayingBar-center {
   margin-top: 28px;
-
-  [data-testid='control-button-playpause'] {
-    z-index: 1;
-  }
-
-  [data-testid='control-button-skip-back'] {
-    margin-right: -42px;
-    padding-right: 32px;
-    border-left: 2px dotted var(--spice-button);
-  }
-
-  [data-testid='control-button-skip-forward'] {
-    margin-left: -42px;
-    padding-left: 32px;
-    border-right: 2px dotted var(--spice-button);
-  }
-
-  :has(> .song-preview-attacher) {
-    transform: translateX(0);
-    transition: transform 0.2s ease-out;
-
-    &:hover {
-      &[data-testid='control-button-skip-back'] {
-        transform: translateX(-3px);
-      }
-
-      &[data-testid='control-button-skip-forward'] {
-        transform: translateX(3px);
-      }
-    }
-  }
-
-  .song-preview-attacher {
-    img,
-    svg {
-      animation: glow-pulse 2.8s infinite ease-in-out;
-    }
-  }
-
-  .prev-attacher {
-    --glow-offset-x: -8px;
-  }
-
-  .next-attacher {
-    --glow-offset-x: 8px;
-  }
 }
 
 .Root__now-playing-bar .playback-bar {
