+++ src/extensions/track-peek/index.css
@@ -1,0 +1,105 @@
+@import '../../styles/animations.css';
+
+.main-nowPlayingBar-center {
+  [data-testid='general-controls'] {
+    container: general-controls / inline-size;
+    position: relative;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .player-controls__left,
+  .player-controls__right {
+    width: 50%;
+    position: absolute;
+  }
+
+  .player-controls__left {
+    right: calc(50% - 8px);
+  }
+
+  .player-controls__right {
+    left: calc(50% - 8px);
+  }
+
+  [data-testid='control-button-playpause'] {
+    z-index: 1;
+  }
+
+  :has([data-testid='control-button-playpause'][aria-label='Play']) {
+    .song-preview-attacher {
+      img,
+      svg {
+        animation: unset;
+      }
+    }
+  }
+
+  [data-testid='control-button-skip-back'] {
+    padding-right: 32px;
+    border-left: 2px dotted var(--spice-button);
+    mask-image: radial-gradient(
+      16px circle at 100% 50%,
+      transparent 99%,
+      #000 100%
+    );
+  }
+
+  [data-testid='control-button-skip-forward'] {
+    padding-left: 32px;
+    border-right: 2px dotted var(--spice-button);
+    mask-image: radial-gradient(
+      16px circle at 0% 50%,
+      transparent 99%,
+      #000 100%
+    );
+  }
+
+  :has(> .song-preview-attacher) {
+    width: unset;
+    flex-basis: 178px;
+    transform: translateX(0);
+    transition: transform 0.2s ease-out;
+
+    &:hover {
+      &[data-testid='control-button-skip-back'] {
+        transform: translateX(-3px);
+      }
+
+      &[data-testid='control-button-skip-forward'] {
+        transform: translateX(3px);
+      }
+    }
+  }
+
+  @container general-controls (max-width: 300px) {
+    :has(> .song-preview-attacher) {
+      flex: unset;
+      background-color: unset;
+      border: unset;
+    }
+
+    .song-preview-attacher {
+      display: none !important;
+    }
+  }
+
+  .song-preview-attacher {
+    img,
+    svg {
+      animation: glow-pulse 2.8s infinite ease-in-out;
+    }
+  }
+
+  .prev-attacher {
+    --glow-offset-x: -8px;
+  }
+
+  .next-attacher {
+    --glow-offset-x: 8px;
+  }
+
+  [data-testid='playback-progressbar'] [data-testid='progress-label'] {
+    z-index: 2;
+  }
+}
