+++ src/extensions/track-peek/index.js
@@ -1,0 +1,5 @@
+import { waitHostReady } from '@/utils/index.js';
+import './index.css';
+
+await waitHostReady();
+import('./root.js');
