+++ src/utils/index.js
@@ -1,0 +1,10 @@
+export async function waitHostReady() {
+  await new Promise((res) => Spicetify.Events.webpackLoaded.on(res));
+  await new Promise((res) => {
+    const checkPlayerAPI = () => {
+      if (Spicetify.Player.origin?._state) return res();
+      setTimeout(checkPlayerAPI, 100);
+    };
+    checkPlayerAPI();
+  });
+}
